LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scrollable-overflow-height-with-flex-item-margin-inline-end-rtl-expected.txt


FAIL Check scrollHeight with overflow: auto assert_equals: expected 950 but got 100
FAIL Check scrollHeight with overflow: scroll assert_equals: expected 950 but got 85
FAIL Check scrollHeight with overflow: hidden assert_equals: expected 950 but got 100
PASS Check scrollHeight with overflow: clip
PASS Check scrollHeight with overflow: visible
PASS Check scrollHeight with overflowX: visible, overflowY: clip
FAIL Check scrollHeight with overflowX: visible, overflowY: hidden assert_equals: expected 950 but got 100
FAIL Check scrollHeight with overflowX: visible, overflowY: auto assert_equals: expected 950 but got 100
